x2cpg: add class for secure xml parsing
useful for queries
