+++ querydb/src/main/scala/io/joern/scanners/android/AndroidMisconfigurations.scala
@@ -21,31 +21,7 @@ object AndroidMisconfigurations extends QueryBundle {
           |""".stripMargin,
       score = 3,
       withStrRep({ cpg =>
-        import javax.xml.parsers.SAXParserFactory
-        import scala.xml.{Elem, XML}
-
-        object SecureXmlParsing {
-          def parseXml(content: String): Option[Elem] = {
-            try {
-              val spf = SAXParserFactory.newInstance()
-
-              spf.setValidating(false)
-              spf.setNamespaceAware(false)
-              spf.setXIncludeAware(false)
-              spf.setFeature("http://xml.org/sax/features/validation", false)
-              spf.setFeature("http://apache.org/xml/features/disallow-doctype-decl", false)
-              spf.setFeature("http://apache.org/xml/features/nonvalidating/load-dtd-grammar", false)
-              spf.setFeature("http://apache.org/xml/features/nonvalidating/load-external-dtd", false)
-              spf.setFeature("http://xml.org/sax/features/external-parameter-entities", false)
-              spf.setFeature("http://xml.org/sax/features/external-general-entities", false)
-
-              Some(XML.withSAXParser(spf.newSAXParser()).loadString(content))
-            } catch {
-              case _: Throwable =>
-                None
-            }
-          }
-        }
+        import io.joern.x2cpg.utils.xml.SecureXmlParsing
 
         val androidUri = "http://schemas.android.com/apk/res/android"
         cpg.configFile
